backend/requirements.txt
Flask
psycopg2-binary
PyJWT
Celery
redis
